mvp/demand/pyproject.toml
[project]
name = "demand-mvp"
version = "0.1.0"
description = "Unified demand MVP for multiple dimensions/facts"
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115.0",
  "uvicorn>=0.30.0",
  "pydantic>=2.8.0",
  "psycopg[binary,pool]>=3.2.0",
  "python-dotenv>=1.0.1",
  "openai>=1.0.0",
  "scikit-learn>=1.3.0",
  "numpy>=1.24.0",
  "pandas>=2.0.0",
  "scipy>=1.11.0",
  "mlflow>=2.8.0",
  "matplotlib>=3.7.0",
  "seaborn>=0.12.0",
  "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
  "pytest>=8.3.0",
]
